+++ packages/mindapp/src/ui/components/Panels.tsx
@@ -14,26 +14,17 @@ export const Panels: Component<FlowPanels> = props => {
   return (
     <>
       <Show when={props.topLeft} keyed>
-        {Panel => {
-          const Component = typeof Panel === 'function' ? <Panel /> : Panel;
-          return <div class='absolute top-4 left-4 z-50'>{Component}</div>;
-        }}
+        {Panel => <div class='absolute top-4 left-4 z-50'>{<Panel />}</div>}
       </Show>
 
       {/* Top-Right Panel */}
       <Show when={props.topRight} keyed>
-        {Panel => {
-          const Component = typeof Panel === 'function' ? <Panel /> : Panel;
-          return <div class='absolute top-4 right-4 z-50'>{Component}</div>;
-        }}
+        {Panel => <div class='absolute top-4 right-4 z-50'>{<Panel />}</div>}
       </Show>
 
       {/* Bottom-Left Panel */}
       <Show when={props.bottomLeft} keyed>
-        {Panel => {
-          const Component = typeof Panel === 'function' ? <Panel /> : Panel;
-          return <div class='absolute bottom-4 left-4 z-50'>{Component}</div>;
-        }}
+        {Panel => <div class='absolute bottom-4 left-4 z-50'>{<Panel />}</div>}
       </Show>
     </>
   );